requirements.txt
absl_py==1.3.0
albumentations==1.3.0
coremltools==6.1
filterpy==1.4.5
Flask==2.0.2
ipython==8.7.0
lap==0.4.0
matplotlib==3.5.0
numpy==1.20.0
onnx==1.12.0
# onnx_graphsurgeon==0.0.1.dev5
onnxruntime==1.13.1
onnxsim==0.4.10
opencv_python==4.6.0.66
openvino==2022.2.0
pafy==0.5.5
pandas==1.4.1
Pillow==9.3.0
psutil==5.8.0
pycocotools==2.0.6
PyYAML==6.0
requests==2.26.0
scikit_image==0.19.3
scipy==1.7.3
seaborn==0.11.2
setuptools==58.1.0
#tensorflow==2.7.0
#tensorflowjs==4.1.0
#tensorrt==8.5.1.7
#tflite_runtime==2.1
thop==0.1.1.post2209072238
torch==1.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.64.0
#nvidia-pyindex
#onnx-graphsurgeon